Update file paths for DWCM model output files to use HOME directory

+++ test_dwcm.py
@@ -1,6 +1,7 @@
 import os, sys
 import numpy as np
 import pickle 
+import datetime as dt
 from collections import defaultdict
 from dcms.models import DCMModel, DECMModel, ADECMModel, DWCMModel
 
@@ -82,23 +83,25 @@ def main():
 
     
     # #### Pytorch, $\theta$
-
-    dwcm_old=DWCMModel(aux[2], aux[3])
-    dwcm_old.solve_tool(tol=1e-5, max_iter=10000, backend='pytorch')
-    # with backend='pytorch'
-    with open('crisis_dwcm_old_theta.pkl', 'wb') as f:
-        pickle.dump(dwcm_old, f)
+    print(f'{dt.datetime.now():%H:%M:%S} DWCM, pytorch, theta')
+    if not os.path.exists(HOME+f'/test/crisis_dwcm_old_theta.pkl'):
+        dwcm_old=DWCMModel(aux[2], aux[3])
+        dwcm_old.solve_tool(tol=1e-5, max_iter=10000, backend='pytorch')
+        # with backend='pytorch'
+        with open(HOME+f'/test/crisis_dwcm_old_theta.pkl', 'wb') as f:
+            pickle.dump(dwcm_old, f)
     
     # #### Pytorch, GS
+    print(f'{dt.datetime.now():%H:%M:%S} DWCM, pytorch, GS')
+    if not os.path.exists(HOME+f'/test/crisis_dwcm_old_gs.pkl'):
+        dwcm_old2=DWCMModel(aux[2], aux[3])
+        dwcm_old2.solve_tool(tol=1e-5, max_iter=10000, backend='pytorch', variant='gauss-seidel')
 
-    dwcm_old2=DWCMModel(aux[2], aux[3])
-    dwcm_old2.solve_tool(tol=1e-5, max_iter=10000, backend='pytorch', variant='gauss-seidel')
-
-    with open('crisis_dwcm_old_gs.pkl', 'wb') as f:
-        pickle.dump(dwcm_old2, f)
+        with open(HOME+f'/test/crisis_dwcm_old_gs.pkl', 'wb') as f:
+            pickle.dump(dwcm_old2, f)
 
     # #### Numba, $\theta$, n_procs=8
-
+    print(f'{dt.datetime.now():%H:%M:%S} DWCM, numba, theta, n_procs=8')
     dwcm=DWCMModel(aux[2], aux[3])
     nprocs=8
     dwcm.solve_tool(tol=1e-5, max_iter=10000, nprocs=nprocs, backend='numba')
@@ -107,6 +110,7 @@ def main():
         pickle.dump(dwcm, f)
 
     # #### Numba, $\theta$, n_procs=1
+    print(f'{dt.datetime.now():%H:%M:%S} DWCM, numba, theta, n_procs=1')
     dwcm=DWCMModel(aux[2], aux[3])
     nprocs=1
     dwcm.solve_tool(tol=1e-5, max_iter=10000, nprocs=nprocs, backend='numba')
@@ -115,6 +119,7 @@ def main():
         pickle.dump(dwcm, f)
 
     # #### Numba, GS, n_procs=8
+    print(f'{dt.datetime.now():%H:%M:%S} DWCM, numba, GS, n_procs=8')
     nprocs=8
     dwcm_gs=DWCMModel(aux[2], aux[3])
     dwcm_gs.solve_tool(tol=1e-5, max_iter=10000, variant='gauss-seidel', nprocs=nprocs, backend='numba')
@@ -123,6 +128,7 @@ def main():
         pickle.dump(dwcm_gs, f)
 
     # #### Numba, GS, n_procs=1
+    print(f'{dt.datetime.now():%H:%M:%S} DWCM, numba, GS, n_procs=1')
     nprocs=1
     dwcm_gs=DWCMModel(aux[2], aux[3])
     dwcm_gs.solve_tool(tol=1e-5, max_iter=10000, variant='gauss-seidel', nprocs=nprocs, backend='numba')
